requirements.txt
absl-py==0.8.1
asgiref==3.2.6
astor==0.8.0
astroid==2.3.2
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.1
cycler==0.10.0
cytoolz==0.10.1
dask==2.9.1
decorator==4.4.1
Django==3.0.4
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.8.0
idna==2.9
imageio==2.6.1
imutils==0.5.3
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.1.2
mccabe==0.6.1
mkl-fft==1.0.6
mkl-random==1.0.1.1
mkl-service==2.3.0
networkx==2.4
numpy==1.17.4
olefile==0.46
opt-einsum==3.1.0
opencv-python==4.2.0.34
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
protobuf==3.11.2
Pygments==2.5.2
pylint==2.4.3
pyparsing==2.4.6
pyreadline==2.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
PyYAML==5.2
pyzmq==18.1.0
requests==2.23.0
scikit-image==0.16.2
scipy==1.3.2
six==1.12.0
spyder-kernels==0.5.2
sqlparse==0.3.1
tensorboard==2.0.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
